remove inline style
