+++ package.json
@@ -10,14 +10,18 @@
     "lint": "eslint ."
   },
   "dependencies": {
-    "@react-native-firebase/app": "^15.7.0",
-    "@react-native-firebase/auth": "^15.7.0",
+    "@react-native-firebase/app": "^16.4.6",
+    "@react-native-firebase/auth": "^16.4.6",
+    "@react-native-firebase/firestore": "^16.4.6",
+    "@react-native-firebase/remote-config": "^16.4.6",
     "@react-navigation/bottom-tabs": "^6.4.0",
     "@react-navigation/native": "^6.0.13",
     "@react-navigation/native-stack": "^6.9.1",
     "@reduxjs/toolkit": "^1.8.6",
+    "pod-install": "^0.1.38",
     "react": "18.1.0",
     "react-native": "0.70.3",
+    "react-native-dropdown-picker": "^5.4.3",
     "react-native-safe-area-context": "^4.4.1",
     "react-native-screens": "^3.18.2",
     "react-native-vector-icons": "^9.2.0",
