+++ src/components/DropDown/styles.js
@@ -1,0 +1,20 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    marginLeft: 120,
+    width: 150,
+  },
+  dropdown: {
+    borderColor: 'rgb(242,242,242)',
+    alignItems: 'center',
+    // height: 20,
+    // width: 120,
+    // margin: 10,
+    backgroundColor: 'rgb(242,242,242)',
+  },
+});
+
+export default styles;
