+++ src/navigate/TabNavigator.js
@@ -1,10 +1,14 @@
 import * as React from 'react';
 import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
 import IconAntDesign from 'react-native-vector-icons/AntDesign';
-import Home from '../screen/tabRoute/Home';
-import Cart from '../screen/tabRoute/Cart';
-import Search from '../screen/tabRoute/Search';
-import Account from '../screen/tabRoute/Account';
+//import Home from '../screen/tabRoute/Home';
+import HomeScreen from '../screen/tabRoute/HomeScreen';
+import CartScreen from '../screen/tabRoute/CartScreen';
+import SearchScreen from '../screen/tabRoute/SearchScreen';
+import AccountScreen from '../screen/tabRoute/AccountScreen';
+//import Cart from '../screen/tabRoute/Cart';
+//import Search from '../screen/tabRoute/Search';
+//import Account from '../screen/tabRoute/Account';
 
 const Tab = createBottomTabNavigator();
 
@@ -13,7 +17,7 @@ const TabNavigator = () => {
     <Tab.Navigator screenOptions={{headerShown: false}}>
       <Tab.Screen
         name="home"
-        component={Home}
+        component={HomeScreen}
         options={{
           tabBarIcon: ({size, color}) => (
             <IconAntDesign name={'home'} color={color} size={size} />
@@ -22,7 +26,7 @@ const TabNavigator = () => {
       />
       <Tab.Screen
         name="Search"
-        component={Search}
+        component={SearchScreen}
         options={{
           tabBarIcon: ({size, color}) => (
             <IconAntDesign name={'search1'} color={color} size={size} />
@@ -31,7 +35,7 @@ const TabNavigator = () => {
       />
       <Tab.Screen
         name="Cart"
-        component={Cart}
+        component={CartScreen}
         options={{
           tabBarIcon: ({size, color}) => (
             <IconAntDesign name={'shoppingcart'} color={color} size={size} />
@@ -40,7 +44,7 @@ const TabNavigator = () => {
       />
       <Tab.Screen
         name="Account"
-        component={Account}
+        component={AccountScreen}
         options={{
           tabBarIcon: ({size, color}) => (
             <IconAntDesign name={'profile'} color={color} size={size} />
