+++ src/screen/authetic/LoginScreen/styles.js
@@ -1,0 +1,135 @@
+import {StyleSheet} from 'react-native';
+import Colors from '../../../constants/Colors';
+
+export const styles = StyleSheet.create({
+  conatiner: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+    backgroundColor: Colors.background,
+  },
+  textDesign: {
+    fontSize: 20,
+    fontWeight: '500',
+    textAlign: 'center',
+    padding: '2%',
+  },
+  top: {
+    bottom: '5%',
+  },
+  icon: {
+    bottom: '620%',
+    right: '44%',
+  },
+  iconTwo: {
+    bottom: '12%',
+    right: '45%',
+  },
+  iconThree: {
+    bottom: '14%',
+    left: '38%',
+  },
+  touchable: {
+    bottom: '400%',
+    color: 'rgb(119,193,78)',
+  },
+  btn: {
+    backgroundColor: 'rgb(112,192,59)',
+    height: 40,
+    width: 100,
+    alignItems: 'center',
+    justifyContent: 'center',
+    top: 1,
+  },
+  or: {
+    fontSize: 20,
+    fontWeight: '400',
+    top: '5%',
+  },
+  text: {
+    top: '10%',
+    fontSize: 23,
+    fontWeight: '500',
+  },
+  position: {
+    right: 100,
+  },
+
+  design: {
+    fontSize: 20,
+    color: 'rgb(255,255,255)',
+    backgroundColor: 'rgb(59,89,152)',
+    paddingTop: 14,
+    paddingLeft: 55,
+
+    left: 15,
+    paddingBottom: 14,
+
+    paddingRight: 30,
+    top: 130,
+    // right:300
+  },
+
+  designThree: {
+    top: 88,
+    height: 30,
+    width: 30,
+    left: 35,
+    // right:40
+  },
+  googlePosition: {
+    top: 28,
+    left: 90,
+  },
+
+  designFour: {
+    fontSize: 18,
+    fontWeight: '700',
+    top: 20,
+    backgroundColor: 'rgb(255,255,255)',
+    paddingTop: 14,
+    paddingLeft: 55,
+
+    borderColor: 'rgb(210,210,210)',
+    borderWidth: 2,
+    paddingBottom: 14,
+
+    paddingRight: 30,
+  },
+  designFive: {
+    bottom: 23,
+    height: 30,
+    width: 30,
+    left: 20,
+  },
+  textField: {
+    fontSize: 20,
+    borderBottomColor: 'black',
+    borderBottomWidth: 1,
+    width: '85%',
+    top: '-2%',
+    padding: '2%',
+    margin: '3%',
+  },
+  log: {
+    bottom: -60,
+    fontSize: 15,
+    fontWeight: 'bold',
+    right: 40,
+  },
+  logtwo: {
+    top: 42,
+    left: 100,
+  },
+  errorMsg: {
+    color: 'red',
+    bottom: '3%',
+    right: '30%',
+    fontSize: 15,
+  },
+  textRegister: {
+    color: 'rgb(112,192,59)',
+    fontSize: 15,
+    fontWeight: 'bold',
+  },
+});
