+++ src/screen/tabRoute/HomeScreen/index.js
@@ -2,7 +2,6 @@ import React from 'react';
 import {
   Text,
   View,
-  StyleSheet,
   SafeAreaView,
   Image,
   FlatList,
@@ -10,8 +9,12 @@ import {
   TouchableOpacity,
 } from 'react-native';
 import Feather from 'react-native-vector-icons/Feather';
+import Ionicons from 'react-native-vector-icons/Ionicons';
 import {useDispatch} from 'react-redux';
-import {addtoCart} from '../../redux/action/action';
+
+import {addtoCart} from '../../../redux/action/action';
+import firestore from '@react-native-firebase/firestore';
+import {styles} from './styles';
 
 const data = [
   {
@@ -19,7 +22,7 @@ const data = [
     name: 'mango',
     Price: 100,
     oldPrice: 200,
-    image: require('../../assests/mango.jpeg'),
+    image: require('../../../assests/mango.jpeg'),
     quantity: 0,
   },
 
@@ -28,7 +31,7 @@ const data = [
     name: 'pineapple',
     Price: 100,
     oldPrice: 200,
-    image: require('../../assests/pineapple.jpeg'),
+    image: require('../../../assests/pineapple.jpeg'),
     quantity: 0,
   },
 
@@ -37,7 +40,7 @@ const data = [
     name: 'grapes',
     Price: 100,
     oldPrice: 200,
-    image: require('../../assests/grapes.png'),
+    image: require('../../../assests/grapes.png'),
     quantity: 0,
   },
 
@@ -46,7 +49,7 @@ const data = [
     name: 'banana',
     Price: 100,
     oldPrice: 200,
-    image: require('../../assests/banana.webp'),
+    image: require('../../../assests/banana.webp'),
     quantity: 0,
   },
 
@@ -55,7 +58,7 @@ const data = [
     name: 'Broccoli',
     Price: 100,
     oldPrice: 200,
-    image: require('../../assests/Broccoli.jpeg'),
+    image: require('../../../assests/Broccoli.jpeg'),
     quantity: 0,
   },
 
@@ -64,7 +67,7 @@ const data = [
     name: 'cabbage',
     Price: 100,
     oldPrice: 200,
-    image: require('../../assests/cabbage.jpeg'),
+    image: require('../../../assests/cabbage.jpeg'),
     quantity: 0,
   },
   {
@@ -72,7 +75,7 @@ const data = [
     name: 'Tomato',
     Price: 100,
     oldPrice: 200,
-    image: require('../../assests/Tomato.webp'),
+    image: require('../../../assests/Tomato.webp'),
     quantity: 0,
   },
   {
@@ -80,12 +83,12 @@ const data = [
     name: 'radish',
     Price: 100,
     oldPrice: 200,
-    image: require('../../assests/radish.jpeg'),
+    image: require('../../../assests/radish.jpeg'),
     quantity: 0,
   },
 ];
 
-const Home = ({navigation}) => {
+const HomeScreen = ({navigation}) => {
   const dispatch = useDispatch();
 
   const deliverData = item => {
@@ -93,6 +96,18 @@ const Home = ({navigation}) => {
 
     dispatch(addtoCart(item));
     //console.log(item.id);
+    firestore()
+      .collection('Cart')
+      .add({
+        name: item.name,
+        id: item.id,
+        price: item.Price,
+        oldPrice: item.oldPrice,
+      })
+      .then(() => {
+        console.log('User added!');
+      });
+    navigation.navigate('Cart');
   };
 
   return (
@@ -100,22 +115,29 @@ const Home = ({navigation}) => {
       <SafeAreaView style={styles.constainer}>
         <View style={styles.innerContainer}>
           <View style={styles.header}>
+            <Feather
+              name={'align-justify'}
+              size={30}
+              style={styles.headerIcon}
+            />
+            <Ionicons
+              name={'notifications-outline'}
+              size={30}
+              style={styles.headerIconTwo}
+            />
             <Text style={styles.headerTxt}>Super Fresh</Text>
           </View>
 
-          <View style={{alignItems: 'center'}}>
-            <View style={[styles.flexStyl, {alignItems: 'center'}]}>
+          <View style={styles.superFresh}>
+            <View style={styles.flexStyl}>
               <Image
                 style={styles.imgStyle}
-                source={require('../../assests/Lettuce.jpeg')}
+                source={require('../../../assests/Lettuce.jpeg')}
               />
               <Text style={styles.headertitle}>Super Fresh</Text>
               <Feather name={'chevron-down'} size={20} />
+              <Feather name={'heart'} size={28} style={styles.heartIcon} />
             </View>
-
-            {/* <View>
-          <Feather name={'heart'} size={20} />
-        </View> */}
           </View>
           <View>
             <FlatList
@@ -140,16 +162,14 @@ const Home = ({navigation}) => {
               horizontal={true}
               renderItem={({item}) => {
                 return (
-                  <View
-                    style={[styles.itemContainer, {flexDirection: 'column'}]}>
-                    <View
-                      style={{width: '100%', height: '100%', borderWidth: 0.5}}>
+                  <View style={styles.itemContainer}>
+                    <View style={styles.innerSection}>
                       <View style={styles.block1}>
                         <Image style={styles.bgimage} source={item.image} />
                       </View>
                       <View style={styles.block1}>
-                        <Text style={{marginHorizontal: 10}}>{item.name}</Text>
-                        <Text style={{marginHorizontal: 10}}>
+                        <Text style={styles.productName}>{item.name}</Text>
+                        <Text style={styles.productPrice}>
                           Rs. {item.price} {item.oldPrice}
                         </Text>
 
@@ -176,16 +196,14 @@ const Home = ({navigation}) => {
               numColumns={'2'}
               renderItem={({item}) => {
                 return (
-                  <View
-                    style={[styles.itemContainer, {flexDirection: 'column'}]}>
-                    <View
-                      style={{width: '100%', height: '100%', borderWidth: 0.5}}>
+                  <View style={styles.itemContainer}>
+                    <View style={styles.innerSection}>
                       <View style={styles.block1}>
                         <Image style={styles.bgimage} source={item.image} />
                       </View>
                       <View style={styles.block1}>
-                        <Text style={{marginHorizontal: 10}}>{item.name}</Text>
-                        <Text style={{marginHorizontal: 10}}>
+                        <Text style={styles.productName}>{item.name}</Text>
+                        <Text style={styles.productPrice}>
                           Rs. {item.price} {item.oldPrice}
                         </Text>
 
@@ -208,81 +226,4 @@ const Home = ({navigation}) => {
     </ScrollView>
   );
 };
-
-const styles = StyleSheet.create({
-  constainer: {
-    width: '100%',
-    height: '100%',
-    backgroundColor: 'white',
-  },
-  innerContainer: {
-    marginHorizontal: 20,
-  },
-  header: {
-    alignItems: 'center',
-  },
-  headerTxt: {
-    fontSize: 18,
-    fontWeight: '500',
-    marginVertical: 20,
-    marginBottom: 30,
-  },
-  flexStyl: {
-    flexDirection: 'row',
-  },
-  imgStyle: {
-    width: 50,
-    height: 50,
-    borderRadius: 25,
-    borderWidth: 1,
-  },
-  headertitle: {
-    fontSize: 16,
-    marginHorizontal: 10,
-  },
-  bgimage: {
-    width: '100%',
-    height: '100%',
-    borderWidth: 1,
-    borderColor: 'lightgray',
-  },
-  horizontalSlidebar: {
-    width: 350,
-    height: 200,
-    marginVertical: 10,
-    marginRight: 10,
-    marginTop: 25,
-  },
-  PopTxt: {
-    fontWeight: '500',
-    fontSize: 20,
-    marginVertical: 10,
-  },
-  itemContainer: {
-    width: 200,
-    height: 300,
-    marginVertical: 10,
-    marginRight: 0,
-    marginTop: 25,
-  },
-  block1: {
-    width: 200,
-    height: '60%',
-  },
-  block2: {
-    width: 200,
-    height: '40',
-  },
-  AtoCBtn: {
-    textAlign: 'center',
-    borderWidth: 1,
-    padding: 15,
-    borderRadius: 5,
-  },
-  touchableArea: {
-    marginHorizontal: 30,
-    marginVertical: 10,
-  },
-});
-
-export default Home;
+export default HomeScreen;
