+++ src/screen/tabRoute/SearchScreen/styles.js
@@ -1,0 +1,45 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  conatiner: {
+    flex: 1,
+    margin: 10,
+    paddingVertical: 20,
+  },
+  design: {
+    fontSize: 20,
+    borderColor: 'black',
+    borderWidth: 1,
+    width: '105%',
+    margin: 10,
+    padding: 10,
+    paddingHorizontal: 30,
+    top: '50%',
+    right: 20,
+    backgroundColor: 'rgb(243,243,243)',
+  },
+  icon: {
+    bottom: '-3%',
+    right: '35%',
+  },
+  item: {
+    flex: 1,
+    flexDirection: 'row',
+    alignItems: 'center',
+    paddingVertical: 13,
+    top: 60,
+    right: 50,
+  },
+
+  subName: {
+    right: '120%',
+    top: '25%',
+  },
+  imageStyle: {
+    height: 100,
+    width: 200,
+  },
+  space: {
+    paddingBottom: 20,
+  },
+});
